#2 => 메인현황판 전체전력사용량 및 검침율 WebFlux로 구현 및 실 DB 데이터 연동 (duration 으로 주기 요청)

+++ Ami/src/main/java/com/cnu/ami/dashboard/models/UseDayHourAllListVO.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 public class UseDayHourAllListVO {
 
 	private int time;
-	private float use;
+	private int use;
 
 }
